Build app and implement Player class
This commit adds required files for creating a project and creates the
Player class with a Header and a cpp file.

The player class has a bunch of methods ranging from spawning the player
to increasing/decressing the health and setting the location and
orientation of the player.

The player class is not called in the main function yet.